-Circle -> radius, area, perimeter (force float) -Rectangle -> width & height ; area and perimeter (force int) -shape_info(shape) -> print area and perimeter

+++ python-abc/task_01_duck_typing.py
@@ -26,10 +26,10 @@ class Circle(Shape):
         self.radius = radius
 
     def area(self):
-        return int(math.pi * self.radius ** 2)
+        return float(math.pi * self.radius ** 2)
 
     def perimeter(self):
-        return int(2 * math.pi * self.radius)
+        return float(2 * math.pi * self.radius)
 
 
 class Rectangle(Shape):
@@ -40,10 +40,10 @@ class Rectangle(Shape):
         self.height = height
 
     def area(self):
-        return self.width * self.height
+        return int(self.width * self.height)
 
     def perimeter(self):
-        return 2 * (self.width + self.height)
+        return int(2 * (self.width + self.height))
 
 
 def shape_info(shape):
